requirements.txt
aerich==0.9.1
aiofiles==24.1.0
aiomysql==0.2.0
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
asyncclick==8.2.2.2
asyncpg==0.30.0
babel==2.17.0
bcrypt==3.2.2
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.1
dictdiffer==0.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
kombu==5.5.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
passlib==1.7.4
pendulum==3.1.0
pillow==11.3.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.2
pypika-tortoise==0.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.38.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
text-unidecode==1.3
tortoise-orm==0.25.1
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.36.0
# uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
